+++ functional-tests/functional/input/src/cpp/Enums.cpp
@@ -19,7 +19,9 @@
 // -------------------------------------------------------------------------------------------------
 
 #include "test/Enums.h"
+#include "test/UseEnumOptionSet.h"
 #include "test/UseEnumWithAlias.h"
+#include <unordered_set>
 
 namespace
 {
@@ -30,7 +32,7 @@ flip_enum( const ::test::Enums::InternalError val )
                ? ::test::Enums::InternalError::ERROR_FATAL
                : ::test::Enums::InternalError::ERROR_NONE;
 }
-}  // namespace
+}
 
 namespace test
 {
@@ -75,5 +77,14 @@ UseEnumWithAlias::compare_to_first(const test::EnumWithAlias input) {
 
 test::EnumWithAlias
 UseEnumWithAlias::get_first() { return test::EnumWithAlias::FIRST; }
+
+// UseEnumOptionSet
+
+std::unordered_set<test::EnumOptionSet, lorem_ipsum::test::hash<test::EnumOptionSet>>
+UseEnumOptionSet::round_trip(
+    const std::unordered_set<test::EnumOptionSet, lorem_ipsum::test::hash<test::EnumOptionSet>>& input
+) {
+    return input;
+}
 
 }